+++ src/main/kotlin/generator/types/U64Descriptor.kt
@@ -3,14 +3,14 @@ package generator.types
 import generator.ASMBuilder
 
 class U64Descriptor : TypeDescriptor() {
-    override fun sizeOf(): UInt = 8u
-    override fun copyTo(offsetTo: UInt, offsetFrom: UInt, asmBuilder: ASMBuilder) {
+    override fun sizeOf(): Int = 8
+    override fun copyTo(offsetTo: Int, offsetFrom: Int, asmBuilder: ASMBuilder) {
         asmBuilder.mov(
             "rax",
-            asmBuilder.pointerWithOffset("rsp", offsetFrom.toInt())
+            asmBuilder.pointerWithOffset("rsp", offsetFrom)
         )
         asmBuilder.mov(
-            asmBuilder.pointerWithOffset("rsp", offsetTo.toInt()),
+            asmBuilder.pointerWithOffset("rsp", offsetTo),
             "rax"
         )
     }
